requirements.txt
# ── Server ───────────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn==0.34.0
click==8.3.1
h11==0.14.0
httptools==0.6.4
httpx==0.28.1
aiosqlite==0.20.0
aiofiles==24.1.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# ── System Tray App ──────────────────────────────────────────────────────────
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.31.0
